requirements.txt
asttokens==2.4.1
backtrader==1.9.78.123
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
eodhd==1.0.30
eventkit==1.0.3
executing==2.1.0
fonttools==4.53.1
frozendict==2.4.4
html5lib==1.1
ib-insync==0.9.86
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.0
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
parso==0.8.4
patsy==0.5.6
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
rich==13.8.1
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.2
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
widgetsnbextension==4.0.13
yfinance==0.2.43
